remove surplus commit file
