+++ util/padding.go
@@ -114,3 +114,13 @@ func Lenf(han string) (l int) {
 	}
 	return
 }
+
+func toHz(han string) string {
+	arg0 := Pad("0", " ", Lenf("0"), 0)
+	arg1 := Pad(han, "", Lenf(han), 0)
+	var un string
+	un += arg0 + " "
+	un += arg1 + " "
+	un += "Standalone Projects"
+	return un
+}
